Fixed a type error

+++ server/src/main/java/com/avalanche/tmcs/job_posting/JobPosting.java
@@ -93,7 +93,7 @@ public class JobPosting {
     }
 
     @NotNull
-    public long getMinMatchedRequiredSkills() {
+    public int getMinMatchedRequiredSkills() {
         return minMatchedRequiredSkills;
     }
